+++ CHANGELOG.md
@@ -8,3 +8,8 @@
 * Add `current`, `home`, `create`, `change` and `remove` calls
 * Add `parent` and `deeper` combinators
 * Add `Read` instances for paths
+
+# 0.1.3
+* Add property tests
+* Rename combinator `<~/>` to `</~>` - absolutizing of homeward path
+* Add new combinator `<~^>` - concatenation of homeward and relative paths